pyproject.toml
[tool.poetry]
name = "idr-analytics"
version = "0.1.0"
description = "IDR 시스템 데이터 분석 AI 에이전트 백엔드"
authors = ["IDR Analytics Team"]
packages = [{ include = "app", from = "idr_analytics" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = { extras = ["standard"], version = "^0.30" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0" }
alembic = "^1.14"
asyncpg = "^0.30"
pydantic-settings = "^2.0"
arq = "^0.26"
pandas = "^2.2"
numpy = "^1.26"
prophet = "^1.1"
statsmodels = "^0.14"
scikit-learn = "^1.5"
langchain = "^0.3"
langchain-anthropic = "^0.3"
anthropic = "^0.40"
httpx = "^0.27"
redis = "^5.0"
python-multipart = "*"
python-jose = { extras = ["cryptography"], version = "*" }
passlib = { extras = ["bcrypt"], version = "*" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
httpx = "^0.27"
ruff = "^0.6"
mypy = "^1.11"
pandas-stubs = "^2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"idr_analytics/app/models/*.py" = ["TCH003"]
"idr_analytics/app/schemas/*.py" = ["TCH003"]
"idr_analytics/alembic/*.py" = ["E501", "TCH"]
"idr_analytics/alembic/versions/*.py" = ["E501", "TCH"]
"idr_analytics/app/core/dependencies.py" = ["B008"]

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
explicit_package_bases = true
mypy_path = "idr_analytics"

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "prophet",
    "sklearn.cluster",
    "sklearn.preprocessing",
    "statsmodels.tsa.arima.model",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["idr_analytics/tests"]
